Changes on the workings of the world.
About the world...
The world is just a *database*, i took some time to understand this, and
this means the world should not care about loading in different threads
but itshould have atomic operations. It is not the world that cares
about the multiple workers, but who access the world.

About chunk loading...
My idea to generate chunks started thinking on the lifetime of the
chunk, which should be something like this:

	void generate_chunk(int x, int y, int z) {
		Chunk *chunk = allocate_chunk(x, y, z);
		make_terrain(chunk);
		generate_neighbours(chunk);
		delete_chunk(chunk);
	}

	void generate_neighbours(Chunk *c) {
		for(int z = -1; z <= 1; z++)
		for(int y = -1; y <= 1; y++)
		for(int x = -1; x <= 1; x++)
		{
			if(position_out_of_bounds(
				x + chunk->x,
				y + chunk->y,
				z + chunk->z
			))
				continue;

			generate_chunk(
				chunk->x + x,
				chunk->y + y,
				chunk->z + z);
		}
	}

And this is the whole lifetime of this chunk. But, as you can see, this is
recursive, and because one chunk requires another chunk to exists, this would
basically generate the entire world, and not loading the neighbours is not
exactly an option because those chunks can generate some structures that might
invade the current chunk being generated, for example: Trees. If we don't, we
might have broken strucutres.
But those neighbours don't need to be fully generated, they only need to be
on a *state* that might generate blocks on the current chunk that is being
fully generated.

So, instead, lets transform this function into a state machine, so we can
control when we should stop.

	void generate_chunk(int x, int y, int z, State target_state) {
		Chunk *chunk = find_or_allocate_chunk(x, y, z);
		switch(chunk->state) {
		case NOT_EXISTING:
			if(target_state <= NOT_EXISTING)
				break;

			chunk->state = ALLOCATED;
			/* fallthrough */
		case ALLOCATED:
			if(target_state <= ALLOCATED)
				break;

			make_terrain(chunk);
			chunk->state = TERRAIN_MADE;
		case TERRAIN_MADE:
			if(target_state <= TERRAIN_MADE)
				break;

			generate_neighbours(chunk);
			chunk->state = MERGED;
			break;
		}
	}

	void generate_neighbours(Chunk *c) {
		for(int z = -1; z <= 1; z++)
		for(int y = -1; y <= 1; y++)
		for(int x = -1; x <= 1; x++)
		{
			generate_chunk(
				chunk->x + x,
				chunk->y + y,
				chunk->z + z,
				TERRAIN_MADE);
		}
	}

Every state must contain "in between" states, for example: GENERATED must
contain a GENERA*TING* state, and this states are going to be used as a
way to "wait" the process to complete, in this case i'm relying more on simple
spinlocks instead of mutexes because they are faster and those states
generally don't take too much time processing it.