tree generation (but broken)
I know the reason why this is broken, once the chunk is already
processed, the data is modified and it becomes no long "available"
to generate a tree.
I think i can save the available surfaces on a bitfield, but even so
we are going to have problems with "available" heights, etc. This commit
is just a scratch to what will become, i'm just commiting now because
there's alot to think about those structure generations, specially
because this will affect other structures that i'm planning to be
naturally generated too.