Cleaning up a little bit
I know why the game is crashing and it has something to do with the
queues. Essentially, what is happening is that the chunks are
being generated and clogging up the workg. This shouldn't be a problem
although slower, but this also makes the generation of those buffers
stop because of the mutex in wg_send(), so the whole game stops, but the
other threads don't stop, keeping sending work to glbufferg, which in
turn crashes the game...

Probably the chunk upload should happen in the worker threads, but...
OpenGL is single threaded and i can't upload from the worker threads nor
create new buffers... Nice...