Changing chunk_gen to not load outside chunks.
This commit does two things:
	1. Make the Graphics chunks a state machine
	2. Don't load the chunk outside the load region

This prevents a buffer overflow that was happening without causing
bugs at the exit of the program and let the graphics process other
chunks that needed to process as well, increasing a little bit the
processing.

Maybe there will be other commits regarding the chunk generation speed
by the graphics chunk workers.